Samlingen

+++ Model/Book.cs
@@ -13,7 +13,7 @@ namespace Wemuda_book_app.Model
         public string? Description { get; set; }
         //public string Authors { get; set; }
         public string? Thumbnail { get; set; }
-        public int? AverageRating { get; set; } 
+        public double? AverageRating { get; set; } 
         public int? RatingCount { get; set; }
     }
 }
